refactor(errors): split tool recovery contract
